impl(bigquery): Add augmented JobConfigurationQuery to Job and ListFormatJob (#11819)

+++ google/cloud/bigquery/v2/minimal/internal/json_utils.cc
@@ -30,8 +30,8 @@ void FromJson(std::chrono::milliseconds& field, nlohmann::json const& j,
 
 void ToJson(std::chrono::milliseconds const& field, nlohmann::json& j,
             char const* name) {
-  j[name] =
-      std::chrono::duration_cast<std::chrono::milliseconds>(field).count();
+  j[name] = static_cast<std::int64_t>(
+      std::chrono::duration_cast<std::chrono::milliseconds>(field).count());
 }
 
 void FromJson(std::chrono::hours& field, nlohmann::json const& j,
@@ -45,7 +45,8 @@ void FromJson(std::chrono::hours& field, nlohmann::json const& j,
 
 void ToJson(std::chrono::hours const& field, nlohmann::json& j,
             char const* name) {
-  j[name] = std::chrono::duration_cast<std::chrono::hours>(field).count();
+  j[name] = static_cast<std::int64_t>(
+      std::chrono::duration_cast<std::chrono::hours>(field).count());
 }
 
 void FromJson(std::chrono::system_clock::time_point& field,
@@ -60,9 +61,10 @@ void FromJson(std::chrono::system_clock::time_point& field,
 
 void ToJson(std::chrono::system_clock::time_point const& field,
             nlohmann::json& j, char const* name) {
-  j[name] = std::chrono::duration_cast<std::chrono::milliseconds>(
-                field - std::chrono::system_clock::from_time_t(0))
-                .count();
+  j[name] = static_cast<std::int64_t>(
+      std::chrono::duration_cast<std::chrono::milliseconds>(
+          field - std::chrono::system_clock::from_time_t(0))
+          .count());
 }
 
 GOOGLE_CLOUD_CPP_INLINE_NAMESPACE_END
